+++ bin/document_store.ml
@@ -6,6 +6,9 @@ type document_info = Document.t * Search_result.t array
 
 type t = {
   all_documents : Document.t String_map.t;
+  file_path_filter_text : string;
+  file_path_filter_re : Re.re option;
+  documents_passing_filter : String_set.t;
   search_exp : Search_exp.t;
   search_exp_text : string;
   search_results : Search_result.t array String_map.t;
@@ -17,6 +20,9 @@ let size (t : t) =
 let empty : t =
   {
     all_documents = String_map.empty;
+    file_path_filter_text = "";
+    file_path_filter_re = None;
+    documents_passing_filter = String_set.empty;
     search_exp = Search_exp.empty;
     search_exp_text = "";
     search_results = String_map.empty;
@@ -32,11 +38,18 @@ let single_out ~path (t : t) =
   | Some doc ->
     let search_results = String_map.find path t.search_results in
     let all_documents = String_map.(add path doc empty) in
+    let documents_passing_filter =
+      if String_set.mem path t.documents_passing_filter then (
+        String_set.(add path empty)
+      ) else (
+        String_set.empty
+      )
+    in
     Some
       {
+        t with
         all_documents;
-        search_exp = t.search_exp;
-        search_exp_text = t.search_exp_text;
+        documents_passing_filter;
         search_results = String_map.(add path search_results empty);
       }
 
@@ -49,6 +62,33 @@ let min_binding (t : t) =
       in
       Some (path, (doc, search_results))
     )
+
+let update_file_path_filter file_path_filter_text (t : t) : t =
+  let documents_passing_filter, file_path_filter_re =
+  if String.length file_path_filter_text = 0 then (
+     (t.all_documents
+  |> String_map.to_seq
+  |> Seq.map fst
+  |> String_set.of_seq,
+  None)
+  ) else (
+    match Misc_utils.compile_glob_re file_path_filter_text with
+    | Some re -> (
+          (t.all_documents
+    |> String_map.to_seq
+    |> Seq.map fst
+          |> Seq.filter (Re.execp re)
+          |> String_set.of_seq,
+          Some re)
+      )
+    | None -> (String_set.empty, None)
+  )
+  in
+  { t with
+  file_path_filter_text;
+  file_path_filter_re;
+  documents_passing_filter;
+    }
 
 let update_search_exp pool stop_signal search_exp_text search_exp (t : t) : t =
   if Search_exp.equal search_exp t.search_exp then (
@@ -104,7 +144,8 @@ let of_seq pool (s : Document.t Seq.t) =
     s
 
 let usable_documents (t : t) : (Document.t * Search_result.t array) array =
-  if Search_exp.is_empty t.search_exp then (
+  if Search_exp.is_empty t.search_exp
+  && String.length t.file_path_filter_text = 0 then (
     t.all_documents
     |> String_map.to_seq
     |> Seq.map (fun (_path, doc) -> (doc, [||]))
@@ -114,11 +155,15 @@ let usable_documents (t : t) : (Document.t * Search_result.t array) array =
       t.all_documents
       |> String_map.to_seq
       |> Seq.filter_map (fun (path, doc) ->
+          if String_set.mem path t.documents_passing_filter then (
           let search_results = String_map.find path t.search_results in
           if Array.length search_results = 0 then
             None
           else
             Some (doc, search_results)
+          ) else (
+            None
+          )
         )
       |> Array.of_seq
     in
@@ -150,6 +195,9 @@ let drop (choice : [ `Single of string | `Usable | `Unusable ]) (t : t) : t =
   match choice with
   | `Single path -> (
       { all_documents = String_map.remove path t.all_documents;
+        file_path_filter_text = t.file_path_filter_text;
+        file_path_filter_re = t.file_path_filter_re;
+        documents_passing_filter = String_set.remove path t.documents_passing_filter;
         search_exp = t.search_exp;
         search_exp_text = t.search_exp_text;
         search_results = String_map.remove path t.search_results;
@@ -162,16 +210,22 @@ let drop (choice : [ `Single of string | `Usable | `Unusable ]) (t : t) : t =
       let document_is_usable path =
         String_set.mem path usable_documents_paths
       in
-      let f : 'a. string -> 'a -> bool =
-        fun path _ ->
+      let f1 path =
           match choice with
           | `Usable -> not (document_is_usable path)
           | `Unusable -> document_is_usable path
           | _ -> failwith "unexpected case"
       in
-      { all_documents = String_map.filter f t.all_documents;
+      let f2 : 'a. string -> 'a -> bool =
+        fun path _ ->
+          f1 path
+      in
+      { all_documents = String_map.filter f2 t.all_documents;
+        file_path_filter_text = t.file_path_filter_text;
+        file_path_filter_re = t.file_path_filter_re;
+        documents_passing_filter = String_set.filter f1 t.documents_passing_filter;
         search_exp = t.search_exp;
         search_exp_text = t.search_exp_text;
-        search_results = String_map.filter f t.search_results;
+        search_results = String_map.filter f2 t.search_results;
       }
     )